requirements.txt
django
djangorestframework
django_filter
pillow
djangorestframework-simplejwt